rebase 15_1_0_pre6
