+++ HeterogeneousCore/SonicCore/BuildFile.xml
@@ -2,7 +2,8 @@
 <use name="FWCore/Concurrency"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
+<use name="FWCore/PluginManager"/>
 <use name="FWCore/Utilities"/>
 <export>
   <lib name="1"/>
-</export>
+</export>i
