+++ HeterogeneousCore/SonicTriton/test/tritonTest_cfg.py
@@ -62,6 +62,15 @@ for im,module in enumerate(options.modules):
                 modelName = cms.string(model),
                 modelVersion = cms.string(""),
                 modelConfigPath = cms.FileInPath("HeterogeneousCore/SonicTriton/data/models/{}/config.pbtxt".format(model)),
+                verbose = cms.untracked.bool(options.verbose or options.verboseClient),
+                useSharedMemory = cms.untracked.bool(not options.noShm),
+                compression = cms.untracked.string(options.compression),
+                Retry = cms.VPSet(
+                  cms.PSet(
+                    retryType = cms.string('RetrySameServerAction'),
+                    allowedTries = cms.untracked.uint32(options.tries)
+                  )
+                )
             )
         )
     )
